Organize custom chain components into custom chain folder
